feat: Creating python lambda for uploading reference data into Aurora (#2)
- Lambda created to upload data from S3 into Aurora
- Pipeline updated to deploy Python Lambdas
