--- .circleci/store.yml
@@ -1,64 +1,0 @@
-# #
-#    Stuff that was removed from config.yml, may be useful later to reintroduce.
-
-#      - run:
-#           name: Testing if environment variables have been set
-#           command: |
-#              if [ -z "${AWS_ACCESS_KEY_ID}" ]; then
-#                 echo "Variable is NOT defined, throwing error and ending build pipeline." 
-#                 exit 1
-#               else
-#                 echo "Variable is set" 
-#               fi
-#              if [ -z "${AWS_SECRET_ACCESS_KEY}" ]; then
-#                 echo "Variable is NOT defined, throwing error and ending build pipeline." 
-#                 exit 1
-#               else
-#                 echo "Variable is set"
-#               fi
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
-#
