+++ .gitignore
@@ -1,15 +1,11 @@
 # package directories
-node_modules
-jspm_packages
+__pycache__
 
 # settings directories
 .vscode
 
 # Serverless directories
 .serverless
-.build
-buckets
-.idea
-.dynamodb
 
 **/coverage/
+node_modules
