+++ src/retrievers/serverless.yml
@@ -1,79 +1,79 @@
 service: reference-data-service-retrievers
 
 provider:
-    name: aws
-    region: eu-west-2
-    runtime: python3.8
-    stage: ${opt:stage, 'test'}
-    stackName: ${self:service}-${self:provider.stage}
-    timeout: 900
-    memorySize: 2048
-    iamRoleStatements:
-        - Effect: 'Allow'
-          Action:
-              - 'ssm:GetParameter'
-          Resource: '*'
-        - Effect: 'Allow'
-          Action:
-              - 's3:PutObject'
-              - 's3:GetObject'
-          Resource:
-              - !Sub arn:aws:s3:::fdbt-csv-reference-data-${self:provider.stage}/*
-              - !Sub arn:aws:s3:::fdbt-xml-reference-data-${self:provider.stage}/*
-              - !Sub arn:aws:s3:::fdbt-tnds-data-zipped-${self:provider.stage}/*
+  name: aws
+  region: eu-west-2
+  runtime: python3.8
+  stage: ${opt:stage, 'test'}
+  stackName: ${self:service}-${self:provider.stage}
+  timeout: 900
+  memorySize: 2048
+  iamRoleStatements:
+    - Effect: "Allow"
+      Action:
+        - "ssm:GetParameter"
+      Resource: "*"
+    - Effect: "Allow"
+      Action:
+        - "s3:PutObject"
+        - "s3:GetObject"
+      Resource:
+        - !Sub arn:aws:s3:::fdbt-csv-reference-data-${self:provider.stage}/*
+        - !Sub arn:aws:s3:::fdbt-xml-reference-data-${self:provider.stage}/*
+        - !Sub arn:aws:s3:::fdbt-tnds-data-zipped-${self:provider.stage}/*
 
 custom:
-    enableSchedule:
-        test: false
-        preprod: true
-        prod: true
+  enableSchedule:
+    test: false
+    preprod: true
+    prod: true
 
 functions:
-    NocRetriever:
-        handler: csv-retriever/main.lambda_handler
-        environment:
-            DATA_URL: https://www.travelinedata.org.uk/wp-content/themes/desktop/nocadvanced_download.php?reportFormat=csvFlatFile&allTable%5B%5D=table_noclines&allTable%5B%5D=table_noc_table&allTable%5B%5D=table_public_name&submit=Submit
-            BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
-            CONTENT_TYPE: text/csv
-        events:
-            - schedule:
-                  rate: cron(0 8 * * ? *)
-                  enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
+  NocRetriever:
+    handler: csv-retriever/main.lambda_handler
+    environment:
+      DATA_URL: https://www.travelinedata.org.uk/wp-content/themes/desktop/nocadvanced_download.php?reportFormat=csvFlatFile&allTable%5B%5D=table_noclines&allTable%5B%5D=table_noc_table&allTable%5B%5D=table_public_name&submit=Submit
+      BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
+      CONTENT_TYPE: text/csv
+    events:
+      - schedule:
+          rate: cron(0 8 * * ? *)
+          enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
 
-    NaptanRetriever:
-        handler: csv-retriever/main.lambda_handler
-        environment:
-            DATA_URL: http://naptan.app.dft.gov.uk/DataRequest/Naptan.ashx?format=csv
-            BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
-            CONTENT_TYPE: text/csv
-            TARGET_FILE: Stops.csv
-        events:
-            - schedule:
-                  rate: cron(0 8 * * ? *)
-                  enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
+  NaptanRetriever:
+    handler: csv-retriever/main.lambda_handler
+    environment:
+      DATA_URL: http://naptan.app.dft.gov.uk/DataRequest/Naptan.ashx?format=csv
+      BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
+      CONTENT_TYPE: text/csv
+      TARGET_FILE: Stops.csv
+    events:
+      - schedule:
+          rate: cron(0 8 * * ? *)
+          enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
 
-    TndsRetriever:
-        handler: tnds-ftp-retriever/main.lambda_handler
-        environment:
-            FTP_HOST: ftp.tnds.basemap.co.uk
-            BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
-            ZIPPED_BUCKET_NAME: fdbt-tnds-data-zipped-${self:provider.stage}
-        events:
-            - schedule:
-                  rate: cron(0 8 * * ? *)
-                  enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
+  TndsRetriever:
+    handler: tnds-ftp-retriever/main.lambda_handler
+    environment:
+      FTP_HOST: ftp.tnds.basemap.co.uk
+      BUCKET_NAME: fdbt-csv-reference-data-${self:provider.stage}
+      ZIPPED_BUCKET_NAME: fdbt-tnds-data-zipped-${self:provider.stage}
+    events:
+      - schedule:
+          rate: cron(0 8 * * ? *)
+          enabled: ${self:custom.enableSchedule.${self:provider.stage}, false}
 
-    TndsUnzipper:
-        handler: tnds-ftp-retriever/unzipper.lambda_handler
-        environment:
-            BUCKET_NAME: fdbt-xml-reference-data-${self:provider.stage}
-        events:
-            - s3: fdbt-tnds-data-zipped-${self:provider.stage}
-              event: s3:ObjectCreated:*
+  TndsUnzipper:
+    handler: tnds-ftp-retriever/unzipper.lambda_handler
+    environment:
+      BUCKET_NAME: fdbt-xml-reference-data-${self:provider.stage}
+    events:
+      - s3: fdbt-tnds-data-zipped-${self:provider.stage}
+        event: s3:ObjectCreated:*
 
 package:
-    exclude:
-        - '*'
-    include:
-        - ./csv-retriever/*
-        - ./tnds-ftp-retriever/*
+  exclude:
+    - "*"
+  include:
+    - ./csv-retriever/*
+    - ./tnds-ftp-retriever/*
