--- src/uploaders/handlers/csv-uploader/requirements.txt
@@ -1,2 +1,0 @@
-pytest==5.4.1
-pymysql==0.9.3
