+++ src/uploaders/requirements.txt
@@ -1,0 +1,2 @@
+pymysql==0.9.3
+untangle==1.1.1
