+++ src/uploaders/serverless.yml
@@ -1,0 +1,58 @@
+service: reference-data-service-uploaders
+
+provider:
+  name: aws
+  runtime: python3.8
+  region: eu-west-2
+  stage: ${opt:stage, 'test'}
+  stackName: ${self:service}-${self:provider.stage}
+  timeout: 30
+  memorySize: 256
+  vpc:
+    securityGroupIds:
+      - Fn::ImportValue: "${self:provider.stage}:RefDataUploaderLambdaSG"
+    subnetIds:
+      - Fn::ImportValue: "${self:provider.stage}:PrivateSubnetA"
+      - Fn::ImportValue: "${self:provider.stage}:PrivateSubnetB"
+  environment:
+    RDS_HOST:
+      Fn::ImportValue: "${self:provider.stage}:RdsHostName"
+  iamRoleStatements:
+    - Effect: "Allow"
+      Action:
+        - "ssm:GetParameter"
+      Resource: "*"
+    - Effect: "Allow"
+      Action:
+        - "s3:GetObject"
+        - "s3:ListBucket"
+      Resource:
+        - arn:aws:s3:::fdbt-csv-reference-data-${self:provider.stage}/*
+        - arn:aws:s3:::fdbt-xml-reference-data-${self:provider.stage}/*
+
+custom:
+  pythonRequirements:
+    dockerizePip: false
+
+plugins:
+  - serverless-python-requirements
+
+functions:
+  CsvUploader:
+    handler: csv-uploader/handler.lambda_handler
+    events:
+      - s3: fdbt-csv-reference-data-${self:provider.stage}
+        event: s3:ObjectCreated:*
+
+  XmlUploader:
+    handler: xml-uploader/handler.xml_uploader_handler
+    events:
+      - s3: fdbt-xml-reference-data-${self:provider.stage}
+        event: s3:ObjectCreated:*
+
+package:
+  include:
+    - ./csv-uploader/*
+    - ./xml-uploader/*
+  exclude:
+    - node_modules/**
